Rename userinfo.js to profile.js.
