+++ lib/strategy.js
@@ -6,7 +6,7 @@ var passport = require('passport-strategy')
   , util = require('util')
   , utils = require('./utils')
   , OAuth2 = require('oauth').OAuth2
-  , UserInfo = require('./userinfo')
+  , Profile = require('./profile')
   , SessionStateStore = require('./state/session')
   , AuthorizationError = require('./errors/authorizationerror')
   , TokenError = require('./errors/tokenerror')
@@ -212,7 +212,7 @@ Strategy.prototype.authenticate = function(req, options) {
               self.success(user, info);
             }
             
-            var idProfile = UserInfo.parse(claims);
+            var idProfile = Profile.parse(claims);
             var profile = {};
             utils.merge(profile, idProfile);
             utils.merge(profile, uiProfile);
@@ -276,7 +276,7 @@ Strategy.prototype.authenticate = function(req, options) {
               return done(new Error('Failed to parse user profile'));
             }
             
-            var profile = UserInfo.parse(json);
+            var profile = Profile.parse(json);
             loaded(profile, json, body);
           });
         }); // self._shouldLoadUserProfile
